+++ tests/integration/test_projectconfigs.py
@@ -93,3 +93,38 @@ def test_dynamic_relays(mini_sentry, relay, caller, projects):
     data = resp.json()
     for p in public_keys:
         assert data["configs"][p] is not None
+
+
+def test_invalid_json(mini_sentry, relay):
+    relay = relay(mini_sentry, wait_healthcheck=True)
+
+    body = "{}"  # missing the required `public_keys` field
+    packed, signature = SecretKey.parse(relay.secret_key).pack(body)
+
+    response = relay.post(
+        "/api/0/relays/projectconfigs/?version=2",
+        data=packed,
+        headers={
+            "X-Sentry-Relay-Id": relay.relay_id,
+            "X-Sentry-Relay-Signature": signature,
+        },
+    )
+
+    assert response.status_code == 400  # Bad Request
+    assert "JSON" in response.text
+
+
+def test_invalid_signature(mini_sentry, relay):
+    relay = relay(mini_sentry, wait_healthcheck=True)
+
+    response = relay.post(
+        "/api/0/relays/projectconfigs/?version=2",
+        data='{"public_keys":[]}',
+        headers={
+            "X-Sentry-Relay-Id": relay.relay_id,
+            "X-Sentry-Relay-Signature": "broken",
+        },
+    )
+
+    assert response.status_code == 401  # Unauthorized
+    assert "signature" in response.text